Fix Locking and overwriting endpoints

See merge request smarte_ladesaeulen/pyocpi!2

+++ ocpi/managers.py
@@ -6,12 +6,13 @@ Created on Thu Apr  1 12:47:48 2021
 @author: maurer
 """
 import requests
-from multiprocessing import Lock
+from multiprocessing import RLock
 import json
 import ocpi.models.credentials as mc
 import logging
 from ocpi.decorators import createOcpiHeader
 import secrets
+import os
 
 log = logging.getLogger('ocpi')
 # sender interface
@@ -165,23 +166,27 @@ class CredentialsManager():
         }
 
 
-lock = Lock()
-
 
 class CredentialsDictMan(CredentialsManager):
+    
+    lock = RLock()
+    
     def __init__(self, credentials_roles: mc.CredentialsRole, url, filename='ocpi_creds.json'):
         self.filename = filename
-        with lock:
-            self.writeJson({})
+        with CredentialsDictMan.lock:
+            if not os.path.isfile(filename):
+                self.writeJson({})
         super().__init__(credentials_roles, url)
 
     def readJson(self):
-        with open(self.filename, 'r') as f:
-            return json.load(f)
+        with CredentialsDictMan.lock:
+            with open(self.filename, 'r') as f:
+                return json.load(f)
 
     def writeJson(self, endpoints):
-        with open(self.filename, 'w') as f:
-            json.dump(endpoints, f, indent=4, sort_keys=False)
+        with CredentialsDictMan.lock:
+            with open(self.filename, 'w') as f:
+                json.dump(endpoints, f, indent=4, sort_keys=False)
 
     def isAuthenticated(self, token):
         return token in self.readJson()
@@ -191,14 +196,14 @@ class CredentialsDictMan(CredentialsManager):
             'client_url': client_url,
             'client_token': client_token,
             'endpoints': endpoint_list or []}
-        with lock:
+        with CredentialsDictMan.lock:
             tokens = self.readJson()
             tokens[token] = data
             self.writeJson(tokens)
         log.info(f'current tokens: {tokens}')
 
     def _deleteToken(self, token):
-        with lock:
+        with CredentialsDictMan.lock:
             tokens = self.readJson()
             tokens.pop(token, None)
             self.writeJson(tokens)
